server/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
PyJWT>=2.8.0
django-cors-headers>=4.3.1
gunicorn>=21.2.0
django-environ>=0.11.2
pytest-django
pytest-cov
requests
bs4
lxml
selenium
ffmpeg
click
rich